- Added some statistic info to output

+++ src/AbstractSilhouetteMethod.hpp
@@ -19,7 +19,7 @@ public:
 
 	virtual void clear() = 0;
 
-	virtual size_t getAccelerationStructureSizeBytes() const = 0;
+	virtual uint64_t getAccelerationStructureSizeBytes() const = 0;
 
 protected:
